requirements.txt
gensim==3.6.0
mecab-python3==1.0.5
numpy==1.22.3
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
unidic-lite==1.0.8
